Initial docker environment for plugin development
Local docker environment via docker-compose.yml file with  bindings for the
various plugin php and dist files.

The stack consists an Nginx proxy, a WordPress/php-fpm container and MySQL.
It will currently create a volumes directory in the project root for MySQL and WordPress mounts.

Logs for Nginx and fpm will be mounts to "logs" directory in the project root.

Both volumes and logs directory have been added to default .gitignore.