Add spl_autoload_register when composer autoload missing.
If there is not vendor/autoload use spl_autoload_register. An alternative is
to add a docker volume. mapping to the wordpress service. eg:

```
- ./vendor:/var/www/html/wp-content/plugins/tenup-scaffold/vendor
```

Note that mapping the vendor directory will cause dev dependencies to be included during development,
prior to packaging composer install / update would need to be run with --no-dev option.