Change to use sass/scss by default.
This change add sass-loader into webpack config, and changes all the css assets to use scss by default.

If css is preferred all that is required is change the file names back to their original names with .css extension (ie: remove prefix from _index.scss files) and update the entry files in webpack settings.

Note that in this configuration it is possible to use both css and sass/scss files at the same time.